requirements.txt
Django==3.1.1
djangorestframework
drf-yasg
channels
asgiref
django-widget-tweaks
Pillow==7.2.0
redis
Celery